+++ app/src/main/java/eu/kanade/tachiyomi/extension/manga/model/MangaLoadResult.kt
@@ -3,5 +3,5 @@ package eu.kanade.tachiyomi.extension.manga.model
 sealed interface MangaLoadResult {
     data class Success(val extension: MangaExtension.Installed) : MangaLoadResult
     data class Untrusted(val extension: MangaExtension.Untrusted) : MangaLoadResult
-    data object Error : MangaLoadResult
+    data class Error(val message: String) : MangaLoadResult
 }